mvp_fastapi/requirements.txt
# Web框架
fastapi>=0.104.0
uvicorn[standard]>=0.24.0

# PPT处理
python-pptx>=0.6.23

# 模板和渲染
jinja2>=3.1.0
reportlab>=4.0.0
pillow>=10.0.0

# LangChain 和 LLM 集成
langchain>=0.0.340
langchain-deepseek>=0.0.1

# 环境变量管理
python-dotenv>=1.0.0

# 其他工具
requests>=2.31.0